Unit Calculator/pyproject.toml
[project]
name = "unit-calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-google-genai>=2.0.11",
    "langgraph>=0.2.74",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
]
